requirements.txt
openai>=1.0.0
anthropic>=0.7.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
tiktoken>=0.5.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
aiofiles>=23.0.0
httpx>=0.24.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0
tenacity>=8.2.0